Add files via upload
When I first listened about DOM it was really fun to play with pokéball colors. This part was developed in 2020 a month before the whole webpage and I always wanted to make them both part of the same project. Decided to do this now even though it's not improved yet.